+++ tests/testthat/test-ArtificialExtinction.R
@@ -8,7 +8,6 @@ dataset[5:6, 4:6] <- '?'
 dataset[6, 3] <- '?'
 
 test_that("Errors are handled", {
-
   expect_error(ArtificialExtinction(dataset, 'A', 'E', replaceAmbiguous = 'INVALID'))
   expect_error(ArtificialExtinction(dataset, 'A', 'E', replaceCoded = 'INVALID'))
 })
@@ -16,13 +15,25 @@ test_that("Errors are handled", {
 test_that("Replacements ok", {
   expect_equivalent(rep('?', 6),
                     ArtificialExtinction(dataset,
-                                         c('A', 'B'), 'E', 'ambig')[1:2, 4:6])
+                                         subject = c('A', 'B'),
+                                         template = 'E',
+                                         replaceAmbiguous = 'ambig')[1:2, 4:6])
 
   expect_equivalent(rep('1', 6),
                     ArtificialExtinction(dataset, c('A', 'B'), 'F',
                                          'unif', sampleFrom = 'A')[1:2, 4:6])
 
   expect_true(all(!'?' == ArtificialExtinction(dataset[-6, ], 1:2, 5, 'freq')[1:2, 4:6]))
+
+
+  expect_true(all(ArtificialExtinction(dataset, subject = 'E', template = 'F',
+                                       replaceAmbiguous = 'binary')['E', 3:6] %in% 0:1))
+
+  expect_equivalent(rep('?', 3),
+                    ArtificialExtinction(dataset, subject = 'E', template = 'F',
+                                         replaceAmbiguous = 'binary',
+                                         replaceAll = FALSE)['E', 4:6])
+
 
   nChar <- 100
   dataset <- rbind(subj = c(rep(2, nChar)),